Define ReqRespBlockResponse in lodestar req resp type
