+++ packages/lodestar/src/chain/chain.ts
@@ -6,7 +6,7 @@ import fs from "fs";
 import {CachedBeaconState, computeStartSlotAtEpoch} from "@chainsafe/lodestar-beacon-state-transition";
 import {IBeaconConfig} from "@chainsafe/lodestar-config";
 import {IForkChoice} from "@chainsafe/lodestar-fork-choice";
-import {allForks, Number64, Root, phase0, Slot, ReqRespBlockResponse} from "@chainsafe/lodestar-types";
+import {allForks, Number64, Root, phase0, Slot} from "@chainsafe/lodestar-types";
 import {ILogger} from "@chainsafe/lodestar-utils";
 import {fromHexString, TreeBacked} from "@chainsafe/ssz";
 import {LightClientUpdater} from "@chainsafe/lodestar-light-client/server";
@@ -218,31 +218,6 @@ export class BeaconChain implements IBeaconChain {
       return null;
     }
     return await this.db.block.get(fromHexString(block.blockRoot));
-  }
-
-  /** Returned blocks have the same ordering as `slots` */
-  async getUnfinalizedBlocksAtSlots(slots: Slot[]): Promise<ReqRespBlockResponse[]> {
-    if (slots.length === 0) {
-      return [];
-    }
-
-    const slotsSet = new Set(slots);
-    const minSlot = Math.min(...slots); // Slots must have length > 0
-    const blockRootsPerSlot = new Map<Slot, Promise<Buffer | null>>();
-
-    // these blocks are on the same chain to head
-    for (const block of this.forkChoice.iterateAncestorBlocks(this.forkChoice.getHeadRoot())) {
-      if (block.slot < minSlot) {
-        break;
-      } else if (slotsSet.has(block.slot)) {
-        blockRootsPerSlot.set(block.slot, this.db.block.getBinary(fromHexString(block.blockRoot)));
-      }
-    }
-
-    const unfinalizedBlocks = await Promise.all(slots.map((slot) => blockRootsPerSlot.get(slot)));
-    return unfinalizedBlocks
-      .map((block, i) => ({bytes: block, slot: slots[i]}))
-      .filter((p2pBlock): p2pBlock is ReqRespBlockResponse => p2pBlock.bytes != null);
   }
 
   async processBlock(block: allForks.SignedBeaconBlock, flags?: PartiallyVerifiedBlockFlags): Promise<void> {
