+++ packages/lodestar/src/chain/interface.ts
@@ -1,4 +1,4 @@
-import {allForks, Number64, Root, phase0, Slot, ReqRespBlockResponse} from "@chainsafe/lodestar-types";
+import {allForks, Number64, Root, phase0, Slot} from "@chainsafe/lodestar-types";
 import {CachedBeaconState} from "@chainsafe/lodestar-beacon-state-transition";
 import {IForkChoice} from "@chainsafe/lodestar-fork-choice";
 import {LightClientUpdater} from "@chainsafe/lodestar-light-client/server";
@@ -83,7 +83,6 @@ export interface IBeaconChain {
    * @param slot
    */
   getCanonicalBlockAtSlot(slot: Slot): Promise<allForks.SignedBeaconBlock | null>;
-  getUnfinalizedBlocksAtSlots(slots: Slot[]): Promise<ReqRespBlockResponse[]>;
 
   /** Process a block until complete */
   processBlock(signedBlock: allForks.SignedBeaconBlock, flags?: PartiallyVerifiedBlockFlags): Promise<void>;
