+++ packages/types/src/types.ts
@@ -1,5 +1,3 @@
-import {Slot} from "./primitive/types";
-
 export * from "./primitive/types";
 export {ts as phase0} from "./phase0";
 export {ts as altair} from "./altair";
@@ -9,10 +7,3 @@ export {ts as allForks} from "./allForks";
 
 /** Common non-spec type to represent roots as strings */
 export type RootHex = string;
-
-/** This type helps response to beacon_block_by_range and beacon_block_by_root more efficiently */
-export type ReqRespBlockResponse = {
-  /** Deserialized data of allForks.SignedBeaconBlock */
-  bytes: Buffer;
-  slot: Slot;
-};
